ADDED magic.h to the game and started implementing iteration 3
Iteration 2 is done and work on iteration 3 has started. Removed vectors
from caves but added vectors for handeling the shopMenu. Also made
various changes in the database
NEXT STEPS:
Continue work on iteration 3 with the intergration of magic.
Add requirements for moves purchasable in the shop
Intergrate types and moves into the enemy fully
Maybe consider adding requirements for caves.